aerticket-booking/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-model-utils==5.0.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
pandas==2.2.3
pdf417==0.8.1
Pillow==11.0.0
python-dotenv==1.0.1
pymongo==4.8.0 
pytz==2024.1
PyYAML==6.0
xmltodict==0.14.2
razorpay==1.4.2
requests==2.32.3
timezonefinder==6.5.7
gunicorn==22.0.0
drf-yasg==1.21.7
pycryptodome==3.22.0
RapidFuzz==3.12.2